Refactor API
This refactor makes the overall skipfield API much more leaner and idiomatic. Instead of using two generic parameters, only one is used instead; the remaining implementation of a skipfield variant then depends solely on the single generic parameter. Possible variants can be found in the `variant` module, which makes it clearer which types are valid skipfield variants and which are not.