Add struct RawIter
The RawIter struct is a non-consuming (borrow) double-ended iterator for the RawSkipfield struct.

The implementation of the RawIter struct is basically equivalent to RawIntoIter, with the exception that the new iter function of the Skipfield trait is used to instantiate it. This allows API users to avoid messing with higher-ranked trait bounds for generic T: Skipfield types.